+++ src/reanimated2/Hooks.js
@@ -10,7 +10,7 @@ import {
   requestFrame,
   getTimestamp,
 } from './core';
-import updateProps from './UpdateProps';
+import updateProps, { updatePropsJestWrapper } from './UpdateProps';
 import { initialUpdaterRun, cancelAnimation } from './animations';
 import { getTag } from './NativeMethods';
 import NativeReanimated from './NativeReanimated';
@@ -48,7 +48,7 @@ export function useEvent(handler, eventNames = [], rebuild = false) {
   return initRef;
 }
 
-export function prepareAnimation(animatedProp, lastAnimation, lastValue) {
+function prepareAnimation(animatedProp, lastAnimation, lastValue) {
   'worklet';
   function prepareAnimation(animatedProp, lastAnimation, lastValue) {
     if (Array.isArray(animatedProp)) {
@@ -104,13 +104,7 @@ export function prepareAnimation(animatedProp, lastAnimation, lastValue) {
   return prepareAnimation(animatedProp, lastAnimation, lastValue);
 }
 
-export function runAnimations(
-  animation,
-  timestamp,
-  key,
-  result,
-  animationsActive
-) {
+function runAnimations(animation, timestamp, key, result, animationsActive) {
   'worklet';
   function runAnimations(animation, timestamp, key, result, animationsActive) {
     if (!animationsActive.value) {
@@ -169,7 +163,7 @@ export function runAnimations(
 }
 
 // TODO: recirsive worklets aren't supported yet
-export function isAnimated(prop) {
+function isAnimated(prop) {
   'worklet';
   function isAnimated(prop) {
     if (Array.isArray(prop)) {
@@ -186,7 +180,7 @@ export function isAnimated(prop) {
   return isAnimated(prop);
 }
 
-export function styleDiff(oldStyle, newStyle) {
+function styleDiff(oldStyle, newStyle) {
   'worklet';
   const diff = {};
   Object.keys(oldStyle).forEach((key) => {
@@ -213,7 +207,7 @@ export function styleDiff(oldStyle, newStyle) {
   return diff;
 }
 
-export const validateAnimatedStyles = (styles) => {
+const validateAnimatedStyles = (styles) => {
   'worklet';
   if (typeof styles !== 'object') {
     throw new Error(
@@ -317,6 +311,110 @@ function styleUpdater(
   }
 }
 
+function jestStyleUpdater(
+  viewDescriptor,
+  updater,
+  state,
+  maybeViewRef,
+  adapters,
+  animationsActive,
+  animatedStyle
+) {
+  'worklet';
+  const animations = state.animations || {};
+  const newValues = updater() || {};
+  const oldValues = state.last;
+
+  // extract animated props
+  let hasAnimations = false;
+  Object.keys(animations).forEach((key) => {
+    const value = newValues[key];
+    if (!isAnimated(value)) {
+      delete animations[key];
+    }
+  });
+  Object.keys(newValues).forEach((key) => {
+    const value = newValues[key];
+    if (isAnimated(value)) {
+      prepareAnimation(value, animations[key], oldValues[key]);
+      animations[key] = value;
+      hasAnimations = true;
+    }
+  });
+
+  function frame(timestamp) {
+    const { animations, last, isAnimationCancelled } = state;
+    if (isAnimationCancelled) {
+      state.isAnimationRunning = false;
+      return;
+    }
+
+    const updates = {};
+    let allFinished = true;
+    Object.keys(animations).forEach((propName) => {
+      const finished = runAnimations(
+        animations[propName],
+        timestamp,
+        propName,
+        updates,
+        animationsActive
+      );
+      if (finished) {
+        last[propName] = updates[propName];
+        delete animations[propName];
+      } else {
+        allFinished = false;
+      }
+    });
+
+    if (Object.keys(updates).length) {
+      updatePropsJestWrapper(
+        viewDescriptor,
+        updates,
+        maybeViewRef,
+        adapters,
+        animatedStyle
+      );
+    }
+
+    if (!allFinished) {
+      requestFrame(frame);
+    } else {
+      state.isAnimationRunning = false;
+    }
+  }
+
+  if (hasAnimations) {
+    state.animations = animations;
+    if (!state.isAnimationRunning) {
+      state.isAnimationCancelled = false;
+      state.isAnimationRunning = true;
+      if (_frameTimestamp) {
+        frame(_frameTimestamp);
+      } else {
+        requestFrame(frame);
+      }
+    }
+  } else {
+    state.isAnimationCancelled = true;
+    state.animations = {};
+  }
+
+  // calculate diff
+  const diff = styleDiff(oldValues, newValues);
+  state.last = Object.assign({}, oldValues, newValues);
+
+  if (Object.keys(diff).length !== 0) {
+    updatePropsJestWrapper(
+      viewDescriptor,
+      diff,
+      maybeViewRef,
+      adapters,
+      animatedStyle
+    );
+  }
+}
+
 export function useAnimatedStyle(updater, dependencies, adapters) {
   const viewDescriptor = useSharedValue({ tag: -1, name: null }, false);
   const initRef = useRef(null);
@@ -325,6 +423,10 @@ export function useAnimatedStyle(updater, dependencies, adapters) {
   adapters = !adapters || Array.isArray(adapters) ? adapters : [adapters];
   const adaptersHash = adapters ? buildWorkletsHash(adapters) : null;
   const animationsActive = useSharedValue(true);
+  let animatedStyle;
+  if (process.env.JEST_WORKER_ID) {
+    animatedStyle = useRef({});
+  }
 
   // build dependencies
   if (!dependencies) {
@@ -347,17 +449,33 @@ export function useAnimatedStyle(updater, dependencies, adapters) {
   const maybeViewRef = NativeReanimated.native ? undefined : viewRef;
 
   useEffect(() => {
-    const fun = () => {
-      'worklet';
-      styleUpdater(
-        viewDescriptor,
-        updater,
-        remoteState,
-        maybeViewRef,
-        adapters,
-        animationsActive
-      );
-    };
+    let fun;
+    if (process.env.JEST_WORKER_ID) {
+      fun = () => {
+        'worklet';
+        jestStyleUpdater(
+          viewDescriptor,
+          updater,
+          remoteState,
+          maybeViewRef,
+          adapters,
+          animationsActive,
+          animatedStyle
+        );
+      };
+    } else {
+      fun = () => {
+        'worklet';
+        styleUpdater(
+          viewDescriptor,
+          updater,
+          remoteState,
+          maybeViewRef,
+          adapters,
+          animationsActive
+        );
+      };
+    }
     const mapperId = startMapper(fun, inputs, []);
     return () => {
       stopMapper(mapperId);
@@ -389,7 +507,11 @@ export function useAnimatedStyle(updater, dependencies, adapters) {
     );
   }
 
-  return { viewDescriptor, initial, viewRef };
+  if (process.env.JEST_WORKER_ID) {
+    return { viewDescriptor, initial, viewRef, animatedStyle };
+  } else {
+    return { viewDescriptor, initial, viewRef };
+  }
 }
 
 // TODO: we should make sure that when useAP is used we are not assigning styles
