--- src/reanimated2/__mocks__/UpdateProps.js
@@ -1,68 +1,0 @@
-/* global _updateProps */
-import { processColor } from '../Colors';
-import { makeShareable } from '../core';
-import { Platform } from 'react-native';
-import { _updatePropsJS } from '../js-reanimated';
-
-// copied from react-native/Libraries/Components/View/ReactNativeStyleAttributes
-const colorProps = [
-  'backgroundColor',
-  'borderBottomColor',
-  'borderColor',
-  'borderLeftColor',
-  'borderRightColor',
-  'borderTopColor',
-  'borderStartColor',
-  'borderEndColor',
-  'color',
-  'shadowColor',
-  'textDecorationColor',
-  'tintColor',
-  'textShadowColor',
-  'overlayColor',
-];
-
-const ColorProperties = makeShareable(colorProps);
-
-export const updateProps = (
-  viewDescriptor,
-  updates,
-  maybeViewRef,
-  adapters,
-  animatedStyle
-) => {
-  'worklet';
-
-  const viewName = viewDescriptor.value.name || 'RCTView';
-
-  if (adapters) {
-    adapters.forEach((adapter) => {
-      adapter(updates);
-    });
-  }
-
-  if (Platform.OS !== 'web') {
-    Object.keys(updates).forEach((key) => {
-      if (ColorProperties.indexOf(key) !== -1) {
-        updates[key] = processColor(updates[key]);
-      }
-    });
-  }
-
-  animatedStyle.current.value = {
-    ...animatedStyle.current.value,
-    ...updates,
-  };
-
-  const updatePropsInternal =
-    typeof _updateProps === 'undefined' ? _updatePropsJS : _updateProps;
-
-  updatePropsInternal(
-    viewDescriptor.value.tag,
-    viewName,
-    updates,
-    maybeViewRef
-  );
-};
-
-export default updateProps;
