+++ src/reanimated2/js-reanimated/JSReanimated.js
@@ -8,6 +8,13 @@ export default class JSReanimated {
   _renderRequested = false;
   _mapperRegistry = new MapperRegistry(this);
   _frames = [];
+  timeProvider = null;
+
+  constructor() {
+    this.timeProvider = process.env.JEST_WORKER_ID
+      ? { now: () => Date.now() }
+      : window.performance;
+  }
 
   pushFrame(frame) {
     this._frames.push(frame);
@@ -15,7 +22,7 @@ export default class JSReanimated {
   }
 
   getTimestamp() {
-    return window.performance.now();
+    return this.timeProvider.now();
   }
 
   maybeRequestRender() {
